refactor(panels): Move addWidget to panels store
