+++ front-src/client/components/App/ContextMenu/AddPanelItem.svelte
@@ -1,10 +1,9 @@
 <script>
   import api from "@/api/panels";
   import { _ } from "@/libs/i18next";
-  import addWidget from "./addWidget";
-  import { setEditMode } from "@/stores/panels";
   import MdAdd from "svelte-icons/md/MdAdd.svelte";
   import MenuItem from "@/components/UI/MenuItem.svelte";
+  import { setEditMode, addWidget } from "@/stores/panels";
 
   async function addPanel() {
     await api.add();
